# Backend/openapi.yml
openapi: 3.0.0
info:
  title: Beacon API
  version: 1.0.0
  description: |
    Backend API for Beacon.

    Most `/api/*` endpoints require a JWT in `Authorization: Bearer <token>`.
    Request validation is enforced via `express-openapi-validator`.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/me/feed:
    get:
      tags: [Auth]
      summary: Get activity feed from followed users
      description: Cursor-paginated feed of pins from users the current user follows
      operationId: getFollowFeed
      parameters:
        - in: query
          name: cursor
          required: false
          schema:
            type: integer
          description: Cursor (pin ID) for pagination
      responses:
        200:
          description: Feed items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pin'
                  nextCursor:
                    type: integer
                    nullable: true
                  hasMore:
                    type: boolean
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/users/{userID}/follow:
    post:
      tags: [Auth]
      summary: Follow a user
      description: Follows the given user. Idempotent — following twice has no effect.
      operationId: followUser
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Now following
        400:
          description: Cannot follow yourself
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Auth]
      summary: Unfollow a user
      description: Removes the follow relationship
      operationId: unfollowUser
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Unfollowed
        404:
          description: Not following this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pin:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the pin
          example: 1
        creatorID:
          type: integer
          description: ID of the account that created the pin
          example: 1
        email:
          type: string
          format: email
          description: Email of the pin creator
          example: "user@example.com"
        latitude:
          type: number
          description: self explanatory
        longitude:
          type: number
          description: self explanatory
        title:
          type: string
          maxLength: 200
          description: Title of the pin
          example: "Golden Gate Overlook"
          nullable: true
        description:
          type: string
          maxLength: 500
          description: Description/message content of the pin
          example: "Welcome to Beacon!"
          nullable: true
        image:
          type: string
          maxLength: 2000
          description: URL of the image
          example: "https://images.unsplash.com/photo-1516192518150-0d8fee5425e3"
          nullable: true
        address:
          type: string
          maxLength: 200
          description: Address of the pin location
          example: "123 Main St, San Francisco, CA"
          nullable: true
        tags:
          type: string
          maxLength: 500
          description: Tags stored as a JSON-encoded string array
          example: '["scenic","viewpoint","nature"]'
          nullable: true
        likes:
          type: integer
          description: Base like count stored on the pin
          example: 0
          nullable: true
        createdAt:
          type: string
          description: Timestamp when the pin was created
          example: "2026-01-17 12:00:00"
          nullable: true
        userStatus:
          type: string
          enum: [visited, wishlist]
          nullable: true
          description: |
            The requesting user's status for this pin, if any. Only populated on
            authenticated endpoints.
        trendingScore:
          type: number
          nullable: true
          description: |
            Computed trending score `likes + 3 * max(0, 1 - ageDays / days)`.
            Only present on `/api/pins/trending`.
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Invalid token"
        error:
          oneOf:
            - type: string
            - type: object
          description: Error identifier/message (some endpoints return 'error' instead of 'message')
          example: "Failed to share itinerary"
        errors:
          type: array
          description: Detailed validation errors (OpenAPI validator)
          items:
            type: object
        details:
          type: string
          description: Additional error details
          example: "FOREIGN KEY constraint failed"
...